Refactor : client/service 내부의 config를 constants/webRTC로 이동 통합
상세:
- 사용하던 코드에서 path도 함께 수정함
